Major changes to takeout_page.html
Takeout page now has a menu, a shopping cart, and a confirmation button that navigates to a new page. The shopping cart is cosmetic because it doesn't actually save any of the data or prices of the food, and only one of each item can be selected at a time, but it shows up on screen and uses javascript.